Tidy current code Rename bracketitem Remove unused functions
