+++ src/engraving/dom/engravingobject.cpp
@@ -31,7 +31,7 @@
 #include "style/textstyle.h"
 #include "types/typesconv.h"
 
-#include "bracketItem.h"
+#include "bracketitem.h"
 #include "linkedobjects.h"
 #include "masterscore.h"
 #include "score.h"
