+++ src/engraving/dom/system.cpp
@@ -34,7 +34,8 @@
 #include "beam.h"
 #include "box.h"
 #include "bracket.h"
-#include "bracketItem.h"
+#include "bracketitem.h"
+#include "chord.h"
 #include "chordrest.h"
 #include "factory.h"
 #include "measure.h"
