+++ src/importexport/guitarpro/internal/importgtp-gp6.cpp
@@ -32,7 +32,7 @@
 #include "engraving/dom/barline.h"
 #include "engraving/dom/box.h"
 #include "engraving/dom/bracket.h"
-#include "engraving/dom/bracketItem.h"
+#include "engraving/dom/bracketitem.h"
 #include "engraving/dom/chord.h"
 #include "engraving/dom/clef.h"
 #include "engraving/dom/dynamic.h"
