+++ src/importexport/musicxml/internal/import/importmusicxmlpass1.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "engraving/dom/box.h"
-#include "engraving/dom/bracketItem.h"
+#include "engraving/dom/bracketitem.h"
 #include "engraving/dom/factory.h"
 #include "engraving/dom/layoutbreak.h"
 #include "engraving/dom/measure.h"
