+++ src/propertiespanel/qml/MuseScore/PropertiesPanel/notation/brackets/bracketsettingsmodel.cpp
@@ -22,7 +22,7 @@
 #include "bracketsettingsmodel.h"
 
 #include "engraving/dom/bracket.h"
-#include "engraving/dom/bracketItem.h"
+#include "engraving/dom/bracketitem.h"
 #include "engraving/dom/score.h"
 #include "engraving/dom/staff.h"
